Added testing framework
There is a super class to steam new test from if we add more, and add in
Lucas's latency test if he is able to give us.
Made server and client load testing ability.
Made a new Dockerfile for the testing this is so when it runs it starts
up a stream, this make the testing easier.